First compilable version
This version of ABS will compile. The server is hardcoded to startup
when you run ABS.java. You can use the configure file to change config
values. There is no sql-file or anything that will set up the database
correctly yet. You can connect to the server from for example a putty
raw-session to the server and run the commands adduser and quit as
client local specefied in the protocol.txt file.